pyproject.toml
[project]
name = "fixed_maze_analyses"
version = "0.1.0"
description = "packages that are used for analyses on the fixed maze"
authors = [
  { name = "Chongyu Qin", email = "chongyu.qin.20@ucl.ac.uk" }
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"

# Dependencies needed to use the package
dependencies = [
    "torch >= 2.7.1",
    "numpy >= 2.3.1",
    "typing-extensions >= 4.14.1",
    "pymongo >= 4.14.0",
    "python-dotenv >= 1.1.1",  # don't forget lmdp package
    "pandas >= 2.3.1",
    "matplotlib >= 3.10.5",
    "scipy >= 1.16.1",
    "scikit-learn >= 1.7.1",
    "statsmodels >= 0.14.5",
    "tqdm >= 4.67.1",
    "tensorboard >= 2.20.0",
    "optuna >= 4.5.0"
]

[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]